Issue: #4042: Split and Organize Checkstyle inputs by Test for MethodLength
