+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/sizes/methodlength/InputSimple.java
@@ -3,7 +3,7 @@
 // Created: Feb-2001
 // Ignore error
 ////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks.sizes;
+package com.puppycrawl.tools.checkstyle.checks.sizes.methodlength;
 import java.io.*;
 /**
  * Contains simple mistakes:
